requirements.txt
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
markdownify==1.1.0
pocketflow==0.0.2
requests==2.32.4
six==1.17.0
soupsieve==2.7
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
PyYAML==6.0.2
